+++ artifacts/@openzeppelin/contracts/token/ERC20/IERC20.sol/IERC20.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "../../../../../build-info/d1ee078b7fa0114610b85668cc7f4548.json"
+  "buildInfo": "../../../../../build-info/e35c60e4ea9b3e089830bd47e366b9b6.json"
 }
